makes a separate file for determinism testing
